Move State views into dedicated file
